Added basic functionality to integrate features/bundles
This improvement is tend to be aimed on flexibility of integrating
modules(features/OSGi bundles) into the platform. End-user should be
capable to use integrated bundle after restarting the application.

DialogUtil is intended to include basic instruments for working with jfx
dialogs

Excluding POM-editing functionality in separate Util can be considered